Move OwnedBufIter to a new file.
